+++ examples/parser.php
@@ -2,11 +2,10 @@
 require_once __DIR__ . '/../src/GlobalException.php';
 require_once __DIR__ . '/../src/CustomizableException/CustomizableException.php';
 require_once __DIR__ . '/../src/CustomizableException/Parser.php';
-require_once __DIR__ . '/../src/Examples/CustomizableException/ExceptionConfig.php';
-require_once __DIR__ . '/../src/Examples/CustomizableException/AException.php';
-require_once __DIR__ . '/../src/Examples/CustomizableException/BException.php';
+require_once __DIR__ . '/resources/CustomizableException/CEExampleConfig.php';
+require_once __DIR__ . '/resources/CustomizableException/CEExampleA.php';
+require_once __DIR__ . '/resources/CustomizableException/CEExampleB.php';
 
-use MagicPush\EnterpriseException\Examples\CustomizableException\ExceptionConfig;
 use MagicPush\EnterpriseException\CustomizableException\Parser;
 
 try {
@@ -14,7 +13,7 @@ try {
     echo "...CustomizableException\ExceptionConfig exceptions with the '$section_to_include' "
         . "section only (short format):\n";
     echo var_export(
-            Parser::parse(ExceptionConfig::class, [], ['class_section_list_in' => [$section_to_include]]),
+            Parser::parse(CEExampleConfig::class, [], ['class_section_list_in' => [$section_to_include]]),
             true
         )
         . "\n";
@@ -22,7 +21,7 @@ try {
     echo "----------\n";
 
     echo 'All ...CustomizableException\ExceptionConfig exceptions (extended format):' . "\n";
-    echo var_export(Parser::parse(ExceptionConfig::class, ['is_extended' => true]), true) . "\n";
+    echo var_export(Parser::parse(CEExampleConfig::class, ['is_extended' => true]), true) . "\n";
 } catch (\Exception $e) {
     echo $e->getMessage() . "\n";
 }