+++ examples/resources/GlobalException/GEExampleC.php
@@ -1,0 +1,10 @@
+<?php
+
+/**
+ * The GlobalException class example.
+ *
+ * @package Examples\GlobalException
+ */
+class GEExampleC extends GEExampleConfig
+{
+}